requirements.txt
newsapi-python
python-dotenv
requests
datetime
pytz
IPython
